WIP: removing the misplaced src/main.rs and starting tests/grids.rs
